requirements.txt
anyio==4.0.0
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorlog==6.7.0
dnspython==2.4.2
fastjsonschema==2.18.0
Flask==2.1.3
Flask-Cors==4.0.0
flask-sock==0.6.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
lxml==4.9.3
MarkupSafe==2.0.1
mypy-extensions==1.0.0
pymongo==4.5.0
python-dotenv==1.0.0
reactpy==1.0.2
requests==2.31.0
simple-websocket==0.10.1
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.1.2
wsproto==1.2.0
anyio==4.0.0
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorlog==6.7.0
dnspython==2.4.2
fastjsonschema==2.18.0
Flask==2.1.3
Flask-Cors==4.0.0
flask-sock==0.6.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
lxml==4.9.3
MarkupSafe==2.0.1
mypy-extensions==1.0.0
packaging==23.1
pymongo==4.5.0
python-dotenv==1.0.0
reactpy==1.0.2
requests==2.31.0
simple-websocket==0.10.1
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.1.2
wsproto==1.2.0
